pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpha-research"
version = "0.1.0"
description = "Multi-agent research and adversarial review system for robotics"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "pymupdf>=1.24",
    "typer>=0.12",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "arxiv>=2.1",
    "anthropic>=0.40",
    "sympy>=1.12",     # scripts/sympy_verify.py
    "scipy>=1.11",     # scripts/audit_stats.py
    # alpha_review is installed separately as an editable dependency:
    #   pip install -e ../alpha_review
    # See guidelines/history/refactor_plan.md Part 0 for setup.
    # Phase 0 of the integrated-state-machine plan removed fastapi,
    # uvicorn, and sse-starlette along with the frontend and api/ layer.
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
alpha-research = "alpha_research.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/alpha_research"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
# Integration tests are OPT-IN (they hit real CLIs / APIs and take minutes).
# Run them with:  pytest -m integration
# Run everything: pytest -m ""
addopts = "-m 'not integration'"
markers = [
    "integration: end-to-end tests that hit real CLIs / APIs (opt-in)",
]
